sakura_core/CKeyWordSetMgr.cpp ファイル内のx64ビルド時の警告を除去

+++ sakura_core/CKeyWordSetMgr.cpp
@@ -313,7 +313,7 @@ int CKeyWordSetMgr::DelKeyWord( int nIdx, int nIdx2 )
 	if( 0 >= m_nKeyWordNumArr[nIdx]	){
 		return 3;	//	登録数が0なら上の条件で引っかかるのでここには来ない？
 	}
-	int nDelKeywordLen = wcslen( m_szKeyWordArr[m_nStartIdx[nIdx] + nIdx2] );
+	size_t nDelKeywordLen = wcslen( m_szKeyWordArr[m_nStartIdx[nIdx] + nIdx2] );
 	int  i;
 	int  endPos = m_nStartIdx[nIdx] + m_nKeyWordNumArr[nIdx] - 1;
 	for( i = m_nStartIdx[nIdx] + nIdx2; i < endPos; ++i ){
@@ -370,7 +370,7 @@ void CKeyWordSetMgr::KeywordMaxLen(int nIdx)
 	int nMaxLen = 0;
 	const int nEnd = m_nStartIdx[nIdx] + m_nKeyWordNumArr[nIdx];
 	for( i = m_nStartIdx[nIdx]; i < nEnd; i++ ){
-		len = wcslen( m_szKeyWordArr[i] );
+		len = static_cast<int>(wcslen( m_szKeyWordArr[i] ));
 		if( nMaxLen < len ){
 			nMaxLen = len;
 		}
@@ -481,7 +481,7 @@ int CKeyWordSetMgr::SetKeyWordArr(
 		const wchar_t* pTop = ptr;	// キーワードの先頭位置を保存
 		while( *ptr != L'\t' && *ptr != L'\0' )
 			++ptr;
-		int kwlen = ptr - pTop;
+		ptrdiff_t kwlen = ptr - pTop;
 		wmemcpy( m_szKeyWordArr[i], pTop, kwlen );
 		m_szKeyWordArr[i][kwlen] = L'\0';
 		++ptr;
@@ -532,7 +532,7 @@ int CKeyWordSetMgr::CleanKeyWords( int nIdx )
 		bool bDelKey = false;	//!< trueなら削除対象
 		// 重複するキーワードか
 		const wchar_t* r = GetKeyWord( nIdx, i + 1 );
-		unsigned int nKeyWordLen = wcslen( p );
+		size_t nKeyWordLen = wcslen( p );
 		if( nKeyWordLen == wcslen( r ) ){
 			if( m_bKEYWORDCASEArr[nIdx] ){
 				if( 0 == wmemcmp( p, r, nKeyWordLen ) ){
